refactor: moving files into src folder
