+++ src/pages/account.vue
@@ -1,4 +1,4 @@
 <template>
-	<div></div>
+	<div class="w-full h-full">account</div>
 </template>
 <script setup></script>
